chore: move playground to app router
